fix(ci): update pnpm-lock.yaml and setup viewport setting
