+++ src/app/layout.js
@@ -1,8 +1,8 @@
 import './globals.css';
+import profileData from '../data/profile.json';
+import socialLinksData from '../data/social-links.json';
 import { Inter, Geist } from 'next/font/google';
 import { NavigationHeader } from '@/components/navigation/Navigation';
-import Breadcrumbs from '@/components/navigation/Breadcrumbs';
-import { House } from 'lucide-react';
 import { cn } from '@/lib/utils';
 
 // Configuration des polices avec next/font
@@ -23,6 +23,66 @@ const geist = Geist({
   variable: '--font-display',
   display: 'swap',
 });
+
+/**
+ * Metadata optimized for Hero7-style landing
+ */
+export const metadata = {
+  title: `${profileData.personal.fullName} - ${profileData.personal.title}`,
+  description: `${profileData.personal.tagline} Portfolio moderne et épuré créé avec Next.js et Shadcn UI.`,
+  keywords:
+    'développeur, software engineer, React, Next.js, Shadcn UI, portfolio moderne, clean design',
+
+  // Multi-langue
+  alternates: {
+    canonical: '/',
+    languages: {
+      'fr-FR': '/',
+      'en-US': '/en',
+    },
+  },
+
+  // Open Graph for social sharing
+  openGraph: {
+    title: `${profileData.personal.fullName} - Portfolio Développeur`,
+    description: profileData.personal.tagline,
+    type: 'website',
+    locale: 'fr_FR',
+  },
+
+  // Twitter Card
+  twitter: {
+    card: 'summary_large_image',
+    title: `${profileData.personal.fullName} - ${profileData.personal.title}`,
+    description: profileData.personal.tagline,
+  },
+
+  // Structured data for better SEO
+  other: {
+    'application/ld+json': JSON.stringify({
+      '@context': 'https://schema.org',
+      '@type': 'Person',
+      name: profileData.personal.fullName,
+      jobTitle: profileData.personal.title,
+      description: profileData.personal.tagline,
+      url: process.env.NEXT_PUBLIC_SITE_URL || 'https://portfolio.dev',
+      workLocation: {
+        '@type': 'Place',
+        name: profileData.personal.location,
+      },
+      knowsAbout: profileData.professional.specializations,
+      sameAs: socialLinksData.socialLinks
+        .filter((link) => link.isProfessional)
+        .map((link) => link.url),
+    }),
+  },
+};
+
+export const viewport = {
+  width: 'device-width',
+  initialScale: 1,
+  maximumScale: 1,
+};
 
 export default function RootLayout({ children }) {
   return (
